query/strategy.rs with common Cells struct (formerly WriteQueryData)
